Agrega columnas con id geográficos asignados a poligonos de provincias o departamentos de Argentina (Augment Argentina's districts polygons id)

Función que permite agregar columnas con ids geográficos para utilización de mapas de Argentina y de departamentos para los 24 distritos sub nacionales para una más sencilla la vinculación con bases de datos de diversas fuentes.

add_geo_codes(data)

Arguments

data
data.frame obtenido con get_geo. Los mapas disponibles se pueden chequear con show_arg_codes.

Value

Los valores por defecto en get_geo son codprov_censo para provincia y coddepto_censo para departamentos, respectivamente. Estos corresponden a la codificación de INDEC. Se puede optar por la nomenclatura de 'indra' - correspondiente a los escrutinios provisorios de elecciones nacionales, con las variantes codprov y coddepto, o la de 'iso' con \*_iso , para el nivel provincial.

Details

Respecto el origen de los datos se puede consultar la documentación de ISO 3166-2 - International Organization for Standardization y del INDEC.

codprov y coddepto son las codificaciones de las bases de datos de INDRA, empresa encargada por muchos años de la tarea del escrutinio provisorio y utilizados en polAr.

Examples

get_geo("ARGENTINA")

#> Simple feature collection with 526 features and 2 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -73.57219 ymin: -55.05744 xmax: -53.64222 ymax: -21.74506
#> geographic CRS: WGS 84
#> # A tibble: 526 x 3
#>    codprov_censo coddepto_censo                                         geometry
#>    <chr>         <chr>                                        <MULTIPOLYGON [°]>
#>  1 02            001            (((-58.37501 -34.57959, -58.36018 -34.5833, -58~
#>  2 02            002            (((-58.40084 -34.57126, -58.37015 -34.57725, -5~
#>  3 02            003            (((-58.39365 -34.60154, -58.39189 -34.62916, -5~
#>  4 02            004            (((-58.35695 -34.6198, -58.34056 -34.62002, -58~
#>  5 06            091            (((-58.01702 -34.7812, -58.14787 -34.88854, -58~
#>  6 74            056            (((-67.18339 -33.11857, -66.80601 -33.10576, -6~
#>  7 02            005            (((-58.41672 -34.59976, -58.41122 -34.63762, -5~
#>  8 02            006            (((-58.45919 -34.6064, -58.43405 -34.60458, -58~
#>  9 06            147            (((-61.53276 -35.38011, -61.28866 -35.57524, -6~
#> 10 02            007            (((-58.46017 -34.6158, -58.45229 -34.63255, -58~
#> # ... with 516 more rows

get_geo("TUCUMAN") %>%
   add_geo_codes()

#> Simple feature collection with 17 features and 8 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -66.18101 ymin: -28.01579 xmax: -64.48315 ymax: -26.06159
#> geographic CRS: WGS 84
#> # A tibble: 17 x 9
#>    codprov_censo coddepto_censo codprov coddepto nomdepto_censo  name_prov
#>    <chr>         <chr>          <chr>   <chr>    <chr>           <chr>    
#>  1 90            007            23      013      BURRUYACU       TUCUMAN  
#>  2 90            014            23      012      CRUZ ALTA       TUCUMAN  
#>  3 90            021            23      005      CHICLIGASTA     TUCUMAN  
#>  4 90            028            23      003      FAMAILLA        TUCUMAN  
#>  5 90            035            23      009      GRANEROS        TUCUMAN  
#>  6 90            042            23      007      JUAN B. ALBERDI TUCUMAN  
#>  7 90            049            23      008      LA COCHA        TUCUMAN  
#>  8 90            056            23      011      LEALES          TUCUMAN  
#>  9 90            063            23      002      LULES           TUCUMAN  
#> 10 90            070            23      004      MONTEROS        TUCUMAN  
#> 11 90            077            23      006      RIO CHICO       TUCUMAN  
#> 12 90            084            23      001      CAPITAL         TUCUMAN  
#> 13 90            091            23      010      SIMOCA          TUCUMAN  
#> 14 90            098            23      017      TAFI DEL VALLE  TUCUMAN  
#> 15 90            105            23      016      TAFI VIEJO      TUCUMAN  
#> 16 90            112            23      014      TRANCAS         TUCUMAN  
#> 17 90            119            23      015      YERBA BUENA     TUCUMAN  
#> # ... with 3 more variables: codprov_iso <chr>, name_iso <chr>,
#> #   geometry <MULTIPOLYGON [°]>