Base de datos de grillas

Grillas de Argentina para usar con facet_geo()

Usage

grillas_geofacet

Format

Una lista que contiene 25 data.frames con 5 variables:

code_provincia

identificación del distrito al que pertenece la grilla

col

LONGITUD relativa de los distritos de segundo nivel en la grilla asemejando la geografía

row

LATITUD relativa de los distritos de segundo nivel en la grilla asemejando la geografía

code

código de identificación de cada uno de los distritos que componen la grilla

name

nombre de cada uno de los distritos que componen la grilla

...