Recodifica id de grillas asignados a provincias o departamentos de Argentina (Recode Argentina's districs id in grids)

Función que permite re codificar etiquetas para utilización de grillas de Argentina y de departamentos para los 24 distritos sub nacionales. Ello permite hacer mas sencilla la vinculación con bases de datos de diversas fuentes.

recode_grid(data, type = NULL)

Arguments

data
data.frame obtenido con get_grid. Las grillas disponibles se pueden chequear con show_arg_codes.

type
la variante del código que se quiere definir para la grilla. Las opciones son 'indra', 'indec' o 'iso'.

Value

Los valores por defecto en get_grid son codprov para provincia y coddepto para departamentos, respectivamente. Estos corresponden a la codificación de los escrutinios provisorios de elecciones nacionales y se etiquetaron como 'indra' . Se puede optar por la nomenclatura de 'indec', con la familia \*_censo, para ambos niveles, o la de 'iso' con \*_iso , para el nivel provincial.

Details

Respecto el origen de los datos se puede consultar la documentación de ISO 3166-2 - International Organization for Standardization y del INDEC.

codprov y coddepto son las codificaciones de las bases de datos de INDRA, empresa encargada por muchos años de la tarea del escrutinio provisorio y utilizados en polAr.

Examples

get_grid("ARGENTINA")

#>    name_provincia row col code                name
#> 1       ARGENTINA   1   1   10               JUJUY
#> 2       ARGENTINA   2   3   09             FORMOSA
#> 3       ARGENTINA   2   5   14            MISIONES
#> 4       ARGENTINA   2   1   17               SALTA
#> 5       ARGENTINA   2   2   23             TUCUMAN
#> 6       ARGENTINA   3   1   03           CATAMARCA
#> 7       ARGENTINA   3   3   06               CHACO
#> 8       ARGENTINA   3   4   05          CORRIENTES
#> 9       ARGENTINA   3   2   22 SANTIAGO DEL ESTERO
#> 10      ARGENTINA   4   2   04             CORDOBA
#> 11      ARGENTINA   4   4   08          ENTRE RIOS
#> 12      ARGENTINA   4   1   12            LA RIOJA
#> 13      ARGENTINA   4   3   21            SANTA FE
#> 14      ARGENTINA   5   3   02        BUENOS AIRES
#> 15      ARGENTINA   5   4   01                CABA
#> 16      ARGENTINA   5   1   18            SAN JUAN
#> 17      ARGENTINA   5   2   19            SAN LUIS
#> 18      ARGENTINA   6   2   11            LA PAMPA
#> 19      ARGENTINA   6   1   13             MENDOZA
#> 20      ARGENTINA   7   1   15             NEUQUEN
#> 21      ARGENTINA   7   2   16           RIO NEGRO
#> 22      ARGENTINA   8   1   07              CHUBUT
#> 23      ARGENTINA   9   1   20          SANTA CRUZ
#> 24      ARGENTINA  10   1   24    TIERRA DEL FUEGO

get_grid("ARGENTINA") %>%
   recode_grid(type = "iso")

#>    row col                name code
#> 1    1   1               JUJUY AR-Y
#> 2    2   3             FORMOSA AR-P
#> 3    2   5            MISIONES AR-N
#> 4    2   1               SALTA AR-A
#> 5    2   2             TUCUMAN AR-T
#> 6    3   1           CATAMARCA AR-K
#> 7    3   3               CHACO AR-H
#> 8    3   4          CORRIENTES AR-W
#> 9    3   2 SANTIAGO DEL ESTERO AR-G
#> 10   4   2             CORDOBA AR-X
#> 11   4   4          ENTRE RIOS AR-E
#> 12   4   1            LA RIOJA AR-F
#> 13   4   3            SANTA FE AR-S
#> 14   5   3        BUENOS AIRES AR-B
#> 15   5   4                CABA AR-C
#> 16   5   1            SAN JUAN AR-J
#> 17   5   2            SAN LUIS AR-D
#> 18   6   2            LA PAMPA AR-L
#> 19   6   1             MENDOZA AR-M
#> 20   7   1             NEUQUEN AR-Q
#> 21   7   2           RIO NEGRO AR-R
#> 22   8   1              CHUBUT AR-U
#> 23   9   1          SANTA CRUZ AR-Z
#> 24  10   1    TIERRA DEL FUEGO AR-V