Agrega capa de Argentina para mapa interactivo (Load Argentina Tiles Template)

Función que descarga capa base de Argentina para mapa interactivo creado con la libreria leaflet.

Usage

addArgTiles(data)

Arguments

data

debe ser un objeto con class "leaflet" "htmlwidget".

Value

Capa base para mapas interactivos

Details

El objetivo es agregar referencias de la base sobre la que se visualizan geometrías de Argentina. Se pueden agregar capas como polígonos descargandolos con get_geo. Los geometrías disponibles se pueden chequear con show_arg_codes.

Examples

 get_geo("TUCUMAN") %>%
 leaflet::leaflet() %>%
 leaflet::addPolygons() %>%
 addArgTiles()